book_warehouse/requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
Django==4.1.1
psycopg2-binary==2.9.3
sqlparse==0.4.3
Pillow==9.2.0
djangorestframework==3.14.0
